LPS-170680 Extract content card to its own component
